Rename integration tests with *TestIT.
